refactor: avoid repetitive fixtures
